+++ backend/app/schemas/budget.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Literal, Optional
 from pydantic import BaseModel, ConfigDict, Field
 
@@ -6,7 +7,7 @@ BudgetStatus = Literal["on_track", "warning", "over_budget"]
 
 class BudgetBase(BaseModel):
     name: str
-    monthly_goal_amount: float = Field(gt=0)
+    monthly_goal_amount: Decimal = Field(gt=0)
 
 class BudgetCreate(BudgetBase):
     month: Optional[int] = Field(default=None, ge=1, le=12)
@@ -14,7 +15,7 @@ class BudgetCreate(BudgetBase):
 
 class BudgetUpdate(BaseModel):
     name: Optional[str] = None
-    monthly_goal_amount: Optional[float] = Field(default=None, gt=0)
+    monthly_goal_amount: Optional[Decimal] = Field(default=None, gt=0)
 
 class BudgetResponse(BudgetBase):
     id: int
@@ -27,7 +28,7 @@ class BudgetResponse(BudgetBase):
     model_config = ConfigDict(from_attributes=True)
 
 class BudgetWithStats(BudgetResponse):
-    spent: float
-    remaining: float
+    spent: Decimal
+    remaining: Decimal
     percent_used: float
     status: BudgetStatus
